requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
certifi==2023.7.22
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
openai==1.2.4
packaging==23.2
parso==0.8.3
platformdirs==4.0.0
prompt-toolkit==3.0.40
psutil==5.9.6
pure-eval==0.2.2
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
wcwidth==0.2.9
